+++ wordpress/docs/CHANGELOG.md
@@ -2,6 +2,16 @@
 
 All notable changes to the **wordpress** extension will be documented in this file.
 
+## [2.102.15] - 2026-05-28
+
+### Fixed
+- run wp codebox fanout concurrently
+
+## [2.102.14] - 2026-05-28
+
+### Fixed
+- apply wp codebox fanout patches
+
 ## [2.102.13] - 2026-05-27
 
 ### Changed